feat - reusable profile and post modal
