--- src/components/Profile/ProfileModal.jsx
@@ -1,55 +1,0 @@
-import { useModal, useProfile } from "contexts"
-import { useEffect, useState } from "react"
-import { ACTION_SET_BIO } from "utils"
-
-const ProfileModal = () => {
-    const [bio, setBio] = useState({})
-    const { modal: { id }, setModal } = useModal()
-    const { getProfileBio, editBio, profileDispatch } = useProfile()
-
-    useEffect(() => {
-        (async () => {
-            const { status, data: { user } } = await getProfileBio(id)
-            if (status === 200) {
-                setBio(user)
-            }
-        })()
-    }, [])
-
-    const handleBioEdit = async () => {
-        const response = await editBio(bio)
-        if (response.status === 201) {
-            profileDispatch({ type: ACTION_SET_BIO, payload: bio })
-        }
-        setModal(m => ({ ...m, type: '', id: '' }))
-        setBio({})
-    }
-
-    return (
-        <section className="flx flx-center modal-container pos-fixed tl-0 z-5">
-            <article className="modal-md flx flx-column bg-primary brd-s pd-s">
-
-                <div className="flx">
-
-                    <input type='text' placeholder='first name' value={bio.firstName} onChange={e => setBio(b => ({ ...b, firstName: e.target.value }))} className='input input-s txt-md pd-xs' />
-
-                    <input type='text' placeholder='last name' value={bio.lastName} onChange={e => setBio(b => ({ ...b, lastName: e.target.value }))} className='input input-s txt-md pd-xs' />
-
-                </div>
-
-                <input type='text' placeholder='portfolio url' value={bio.portfolio ?? ''} onChange={e => setBio(b => ({ ...b, portfolio: e.target.value }))} className='input input-s txt-md pd-xs' />
-
-                <input type='text' placeholder='city' value={bio.city ?? ''} onChange={e => setBio(b => ({ ...b, city: e.target.value }))} className='input input-s txt-md pd-xs' />
-
-                <div onClick={handleBioEdit} className="flx flx-maj-end">
-                    <button className="btn-solid bg-secondary txt-secondary txt-md txt-lcase pd-xs brd-s">
-                        edit
-                    </button>
-                </div>
-
-            </article>
-        </section>
-    )
-}
-
-export default ProfileModal
