+++ src/components/Reusable/index.js
@@ -4,4 +4,4 @@ export { default as FeedPageWrapper } from './FeedPage'
 export { default as Post } from './Post'
 export { default as CreatePost } from './CreatePost'
 export { default as PageHeading } from './PageHeading'
-export { default as Modal } from './Modal'
+export { default as ModalWrapper } from './Modal'
